requirements.txt
python-docx>=0.8.11
cohere>=5.6.0
tkinterdnd2>=0.3.0